fix: update server CDP test imports after cdp-protocol move
Point remaining server unit and integration specs at the relocated
cdp-protocol modules so CI can load them after the directory move.
