+++ packages/server/test/unit/browsers/browser-cri-client_spec.ts
@@ -1,5 +1,5 @@
 import { BrowserCriClient } from '../../../lib/browsers/browser-cri-client'
-import { CriClient } from '../../../lib/browsers/cri-client'
+import { CriClient } from '../../../lib/browsers/cdp-protocol/cri-client'
 import { expect, proxyquire, sinon } from '../../spec_helper'
 import * as protocol from '../../../lib/browsers/protocol'
 import { stripAnsi } from '@packages/errors'
