+++ packages/server/test/unit/browsers/cdp-connection_spec.ts
@@ -1,9 +1,9 @@
 import type CDP from 'chrome-remote-interface'
-import type { CdpClient, CDPConnection, CDPConnectionOptions } from '../../../lib/browsers/cdp-connection'
-import type { debugCdpConnection } from '../../../lib/browsers/debug-cdp-connection'
-import type { CdpEvent, CdpCommand } from '../../../lib/browsers/cdp_automation'
+import type { CdpClient, CDPConnection, CDPConnectionOptions } from '../../../lib/browsers/cdp-protocol/cdp-connection'
+import type { debugCdpConnection } from '../../../lib/browsers/cdp-protocol/debug-cdp-connection'
+import type { CdpEvent, CdpCommand } from '../../../lib/browsers/cdp-protocol/cdp_automation'
 import type ProtocolMapping from 'devtools-protocol/types/protocol-mapping'
-import { CDPTerminatedError, CDPAlreadyConnectedError, CDPDisconnectedError } from '../../../lib/browsers/cri-errors'
+import { CDPTerminatedError, CDPAlreadyConnectedError, CDPDisconnectedError } from '../../../lib/browsers/cdp-protocol/cri-errors'
 import WebSocket from 'ws'
 import pDefer, { DeferredPromise } from 'p-defer'
 const { expect, proxyquire, sinon } = require('../../spec_helper')
@@ -46,7 +46,7 @@ describe('CDPConnection', () => {
 
     CDPImport = sinon.stub()
 
-    CDPConnection = proxyquire('../lib/browsers/cdp-connection', {
+    CDPConnection = proxyquire('../lib/browsers/cdp-protocol/cdp-connection', {
       'chrome-remote-interface': CDPImport,
       './debug-cdp-connection': stubbedDebugger,
     }).CDPConnection
