+++ packages/server/test/unit/browsers/electron_spec.ts
@@ -7,7 +7,7 @@ import { Automation } from '../../../lib/automation'
 import { BrowserCriClient } from '../../../lib/browsers/browser-cri-client'
 import * as electronApp from '../../../lib/util/electron-app'
 import utils from '../../../lib/browsers/utils'
-import { screencastOpts } from '../../../lib/browsers/cdp_automation'
+import { screencastOpts } from '../../../lib/browsers/cdp-protocol/cdp_automation'
 import menu from '../../../lib/gui/menu'
 import * as plugins from '../../../lib/plugins'
 import * as Windows from '../../../lib/gui/windows'
